+++ offer-manager-app/src/main/groovy/life/qbic/portal/offermanager/components/offer/create/CreateOfferViewModel.groovy
@@ -38,7 +38,7 @@ class CreateOfferViewModel {
 
     @Bindable OfferId offerId
     @Bindable String projectTitle
-    @Bindable String projectDescription
+    @Bindable String projectObjective
     @Bindable Customer customer
     @Bindable Affiliation customerAffiliation
     @Bindable ProjectManager projectManager